Organize Footer files in directory.
