+++ src/components/Footer/Footer.js
@@ -1,10 +1,10 @@
 import React from "react";
 import { Link } from "gatsby";
 
-import logo from "../img/logo.svg";
-import facebook from "../img/social/facebook.svg";
-import instagram from "../img/social/instagram.svg";
-import twitter from "../img/social/twitter.svg";
+import logo from "../../img/logo.svg";
+import facebook from "../../img/social/facebook.svg";
+import instagram from "../../img/social/instagram.svg";
+import twitter from "../../img/social/twitter.svg";
 
 import "./Footer.scss";
